### 2022-02-02 #4123 @whedon
[PRE REVIEW]: Lexedata: A Python toolbox to edit CLDF lexical datasets
**Submitting author:** @Anaphory (<a href="http://orcid.org/0000-0002-8155-9089">Gereon Kaiping</a>)
**Repository:** <a href="https://github.com/Anaphory/lexedata/" target ="_blank">https://github.com/Anaphory/lexedata/</a>
**Version:** v1.0.0
