### 2022-02-02 #4123 @Kevin-Mattheus-Moerman
@Anaphory It looks like you you should use `"2,3"` instead of `2,3` for the affiliations for `Natalia Chousou-Polydouri`. You can call the following here to update the paper:
`@whedon generate pdf from branch joss-paper`
