### 2022-02-04 #4123 @Bisaloo
Hi @lmaurits @xrotwang :wave: :wave:, you have been suggested to review [this paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04123/joss.04123/10.21105.joss.04123.pdf) and [python library](https://github.com/Anaphory/lexedata/) for JOSS. Looking at your work and GitHub profile, I agree you are a very good fit.

Would you be available to review this submission?
