### 2022-02-04 #4123 @Bisaloo
@xrotwang, amazing! Thanks for your help. Let me find a second reviewer and we can get started. I'll post instructions then but if you're already curious, you can read more about our process in our [online docs](https://joss.readthedocs.io/en/latest/index.html).
