### 2022-02-04 Anaphory/lexedata#295 @Anaphory
importer.excel_long_format requires metadata file
As @xrotwang pointed out:

To bootstrap a dataset it's unhandy that excel_long_format requires a metadata file, a LanguageTable, and a ParameterTable.
It's useful that it supports those if present, so if language or concept metadata are listed in separate columns from the forms' language references, they can be checked and imported. But the script should work without one.
