### 2022-02-05 Anaphory/lexedata#296 @xrotwang
command in "Tour of lexedata" leads to "empty" screen
The command
```
$ python -c 'from openpyxl import load_workbook
> for row in load_workbook("bantu.xlsx").active.iter_rows():
>   print(*[c.value or "" for c in row], sep="\t")'
[...]
```
print each row of the excel sheet. But since the last couple of rows are empty, I end up with an "empty" screen and have to scroll to see what the document tells me I would see.
