### 2022-02-05 Anaphory/lexedata#297 @xrotwang
Error in lexedata.edit.clean_forms
Running `python -m lexedata.edit.clean_forms` results in
```
$ python -m lexedata.edit.clean_forms
Traceback (most recent call last):
  File "/usr/lib/python3.8/runpy.py", line 194, in _run_module_as_main
    return _run_code(code, main_globals, None,
  File "/usr/lib/python3.8/runpy.py", line 87, in _run_code
    exec(code, run_globals)
  File "/home/robert_forkel/projects/lexedata/lexedata/src/lexedata/edit/clean_forms.py", line 250, in <module>
    forms = list(
  File "/home/robert_forkel/projects/lexedata/lexedata/src/lexedata/edit/clean_forms.py", line 127, in treat_brackets
    yield row | {
TypeError: unsupported operand type(s) for |: 'collections.OrderedDict' and 'dict'
```
