### 2022-02-05 Anaphory/lexedata#297 @xrotwang
It appears that `|` for `dict` was just added in Python 3.9, see https://www.python.org/dev/peps/pep-0584/
