### 2022-02-05 Anaphory/lexedata#297 @xrotwang
Given that test coverage for `lexedata` isn't particularly high (`lexedata.edit.clean_forms` isn't even imported in any test), I'd see three options:
- increase test coverage considerably
- only claim whatever the developers version of Python is as minimal supported version
- switch to python 3.8 for development.

Otherwise I don't think claiming support of python 3.8 would be justified.
