### 2022-02-05 Anaphory/lexedata#297 @xrotwang
I should note that I'd consider python 3.8 support somewhat important. I'm on Ubuntu 20.04 - which is still supported for another 2 years at least - and python 3.8 is the default python 3 installation :)
