### 2022-02-05 Anaphory/lexedata#297 @xrotwang
Oh, and I just read

> Python 3.9+ cannot be used on Microsoft Windows 7 or earlier versions. Last Windows 7 installable version is python 3.8 series.

which - if true - would also speak for maintaining py3.8 compatibility.
