### 2022-02-05 Anaphory/lexedata#298 @xrotwang
Add a CONTRIBUTING.md
Such a document could describe
- how to install the package for development
- how to run the test suite
- how to write tests
