### 2022-02-07 29795f5
Add CONTRIBUTING.md
Fixes #298
