### 2022-02-07 2a1be9f
Fix example spreadsheet output in the tour
Fixes #296
