### 2022-02-07 Anaphory/lexedata#297 @Anaphory
Yes, 3.8 support is the intention, and we are not at good at writing regression tests, otherwise that particular `|` operator would not have made it into release – we did see it being an issue in 3.8 but somehow it slipped back in. I'll drop my development version to 3.8 and see how I can get this replaced and tested.
