### 2022-02-07 1e56b02
Remove dictionary union operator, which is only in 3.9
