+++ src/lexedata/edit/clean_forms.py
@@ -123,13 +123,17 @@ def treat_brackets(
                 )
                 variants.extend(new_variants)
                 comment.extend(new_comments)
-            # TODO: Should we have a message here?
-            yield row | {
+            # We need to avoid dict.update() here, so that in a recursive call
+            # where an earier bracket has already succeeded, still the whole
+            # form cell is skipped.
+            yield {
+                **row,
                 form_column_name: form,
                 variants_column_name: variants,
                 comment_column_name: "; ".join(comment),
             }
         except Skip as e:
+            # TODO: Should we have a message here?
             logger.error(
                 "Line %d: Form '%s' has %s. I did not modify the row.",
                 r,
