### 2022-02-07 Anaphory/lexedata#297 @Anaphory
There is a doctest, though, and 3.8 is in our Github workflow matrix, so I'm a bit confused why this was not picked up by the CI.
