### 2022-02-07 b54502f
Cannot strip None
