### 2022-02-07 Anaphory/lexedata#297 @xrotwang
How are tests supposed to be run?
