### 2022-02-07 Anaphory/lexedata#297 @xrotwang
Just looked at tox.ini. That's quite a bit of machinery :)
