### 2022-02-07 Anaphory/lexedata#297 @xrotwang
So which doctest should run `clean_forms`?
