### 2022-02-07 Anaphory/lexedata#297 @xrotwang
Ok, this error **is** picked up when I run tox locally with py3.8.
