### 2022-02-07 Anaphory/lexedata#297 @xrotwang
From my experience, relying on GitHub actions alone for cross-form testing isn't ideal. So I typically run tox with a couple of pythons locally before release commits.
