### 2022-02-07 Anaphory/lexedata#297 @Anaphory
Well, the actual error you encountered is inside `treat_brackets`, which has its own doctest in its own docstring. Which is designed to encounter both the place where the `|` was and also check that the row is not mutated if there is a skip later. And the CI report https://github.com/Anaphory/lexedata/runs/5096896551?check_suite_focus=true#step:7:96 actually contains three “passing” dots for three doctests inside clean_forms.

Maybe I misconfigured the Github test matrix.

I'll keep the advice to test with multiple versions locally in mind for next release.
