### 2022-02-07 Anaphory/lexedata#297 @xrotwang
It seems you are not **really** running the tests on py3.8. Looking at the logs, it says
> uccessfully setup CPython (3.10.2)
