### 2022-02-07 Anaphory/lexedata#297 @xrotwang
You are also not running on "3.1" as the logs seems to say, but on 3.10 again. To specify 3.10, I needed to use `"3.10"`, i.e. a string, because `3.10` is read in yaml as `3.1`. So, something with your gh actions setup seems wrong.
