### 2022-02-07 Anaphory/lexedata#297 @xrotwang
FWIW, the matrix setup here works: https://github.com/glottolog/pyglottolog/blob/master/.github/workflows/python-package.yml
