+++ .github/workflows/python-tox.yml
@@ -9,16 +9,18 @@ on:
 
 jobs:
   build:
-
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python: [3.8, 3.9, 3.10]
-        pycldf: [1.24.0, 1.21.0, NEWEST]
+        python: ["3.8", "3.9", "3.10"]
+        pycldf: ["1.24.0", "1.21.0", "NEWEST"]
     steps:
     - uses: actions/checkout@v2
     - name: Set up Python ${{ matrix.python }}
       uses: actions/setup-python@v2
+      with:
+        python-version: ${{ matrix.python }}
+        architecture: x64
     - name: Install pycldf ${{ matrix.pycldf }}
       run: |
         python -m pip install --upgrade pip
