### 2022-02-07 Anaphory/lexedata#297 @Anaphory
Yes, I was missing the `with:` clause. I had noticed the `3.1` but attributed it to the output instead of going to the root cause.
