### 2022-02-07 e082c2f
Expand comment on the implementation choice.
