+++ src/lexedata/edit/clean_forms.py
@@ -123,9 +123,10 @@ def treat_brackets(
                 )
                 variants.extend(new_variants)
                 comment.extend(new_comments)
-            # We need to avoid dict.update() here, so that in a recursive call
-            # where an earier bracket has already succeeded, still the whole
-            # form cell is skipped.
+            # We avoid dict.update() here, so that in a recursive call where an
+            # earier bracket has already succeeded, still the whole form cell
+            # is skipped. Or at least I thought that was the logic, except
+            # there are no recursive calls to treat_brackets.
             yield {
                 **row,
                 form_column_name: form,
