### 2022-02-08 #4123 @Bisaloo
Hi @peterdekker :wave:, the authors of this submission identified you as a potential review for this submission, would you be able to review this [paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04123/joss.04123/10.21105.joss.04123.pdf) and [python library](https://github.com/Anaphory/lexedata/) for [JOSS](https://joss.readthedocs.io/en/latest/)?
