### 2022-02-08 709b938
Install different python versions for tests
