### 2022-02-08 7d6feed
Fix typo
Pointed out in https://github.com/Anaphory/lexedata/pull/299#discussion_r800933787
