### 2022-02-08 #4140 @Bisaloo
@xrotwang, I saw you already opened some issues in the repository. Thanks for your efficiency! :rocket:

I'll link the issues here for future reference:

https://github.com/Anaphory/lexedata/issues/296
https://github.com/Anaphory/lexedata/issues/297
https://github.com/Anaphory/lexedata/issues/298
