### 2022-02-08 Anaphory/lexedata#295 @Anaphory
This issue is thus part of the [JOSS review](openjournals/joss-reviews#4140)
