### 2022-02-08 Anaphory/lexedata#301 @xrotwang
Describe how lexedata relates to cldfbench
I think it would be good to mention somewhere in the docs how (the CLDF creation functionality of) lexedata relates to [cldfbench](https://github.com/cldf/cldfbench). From my perspective, both tools have their niche, but it may help describing these respective niches. The way I look at it, `cldfbench` is for cases where
- writing some python code is ok for the data curators
- the raw or source data may be fully beyond the control of CLDF data curators.

Thus, `cldfbench` is supposed to operate under somewhat "extreme" conditions.

`lexedata` chooses a different trade-off: No python required for users, but cooperation between data collectors and data curators is needed. The observation that the conditions under which `lexedata` works are a lot more common in the targeted research community constitutes its "need", I guess.

openjournals/joss-reviews#4140
