### 2022-02-08 bf2abc5
allow contributors to run tox without any errors (which may be caused by trying to upload coverage data to codecov)
