+++ tox.ini
@@ -42,4 +42,3 @@ commands =
   pytest --cov=lexedata --cov-append test/
   python -m scriptdoctest -oFAIL_FAST -oREPORT_UDIFF docs/tour.rst
   python test/check_all_versions_match.py
-  codecov
