+++ .github/workflows/python-tox.yml
@@ -24,7 +24,7 @@ jobs:
     - name: Install pycldf ${{ matrix.pycldf }}
       run: |
         python -m pip install --upgrade pip
-        pip install flake8 pytest tox
+        pip install flake8 pytest tox codecov
         if [ x${{ matrix.pycldf }} == xNEWEST ]
         then
           pip install pycldf
