+++ tox.ini
@@ -23,7 +23,6 @@ extend_select =
 passenv = CI
 
 deps =
-  codecov
   black ~= 21.6b0
   flake8
   flake8-breakpoint
