### 2022-02-14 Anaphory/lexedata#305 @peterdekker
Role of contributors
(for review JOSS [#4140](https://github.com/openjournals/joss-reviews/issues/4140))

I see there is a contributor of a small amount of code in the repository, Walter Fuchs, who is not an author on the JOSS paper. Could you explain what was their role in the project? Thanks!
