### 2022-02-14 Anaphory/lexedata#306 @peterdekker
Installation instructions
https://github.com/openjournals/joss-reviews/issues/4140

I think it would be good to add some instructions to the README file on how to install the package and its requirements, and how to generate the docs. Now I had to guess/google this, I tried ``sudo python3 setup.py install`` and ``make doc``.
