### 2022-02-14 Anaphory/lexedata#306 @xrotwang
Maybe a link to pypi in the README would make it clear that the package can be installed from there. For development installs, the info in CONTRIBUTING.md looks sufficient to me.

(Btw.: `sudo python3 setup.py install` is not recommended, since it may screw up the global python3 installation. Virtual environments are the safer option.)
