### 2022-02-14 Anaphory/lexedata#306 @peterdekker
I see now the installation instructions in the documentation: https://lexedata.readthedocs.io/en/latest/installation.html
There is a general link to the docs in the README, but maybe it would be good to add a link specifically to the Installation page in README, and maybe also a link to CONTRIBUTING.

Also, in that installation page in the docs, it described installing a stable version which is available in pip (``pip install lexedata``), but wouldn't it also be interesting to describe installation of a (possibly newer) clone of the repository (``pip install .``)?
