### 2022-02-14 Anaphory/lexedata#306 @xrotwang
I think installing an unreleased version from a clone should be limited to development. Otherwise things like error reporting might get quite a bit more confusing. Assuming a "release early, release often" release policy, this wouldn't hide much functionality from non-developers, I think.
