### 2022-02-14 Anaphory/lexedata#307 @peterdekker
Permission error when installing development installation in venv
https://github.com/openjournals/joss-reviews/issues/4140

I followed the instructions in CONTRIBUTING to install a development installation of lexedata. I used `venv` to create the virtual environment instead of `conda`.

In a clone of lexedata 1.0.0, I create a venv and run pip install as described in the instructions:
``python3 -m venv lenv
source lenv/bin/activate
pip install -e .[dev,test]``

Output:
```
Obtaining file:///home/peter/git/lexedata-1.0.0
  Installing build dependencies ... done
  Getting requirements to build wheel ... error
   command: /home/peter/git/lexedata-1.0.0/lenv/bin/python3 /tmp/tmp7dbr261c_in_process.py get_requires_for_build_wheel /tmp/tmpx9nxe50i
       cwd: /home/peter/git/lexedata-1.0.0
  Complete output (2 lines):
  running egg_info
  error: [Errno 13] Permission denied
WARNING: Discarding file:///home/peter/git/lexedata-1.0.0. Command errored out with exit status 1: /home/peter/git/lexedata-1.0.0/lenv/bin/python3 /tmp/tmp7dbr261c_in_process.py get_requires_for_build_wheel /tmp/tmpx9nxe50i Check the logs for full command output.
ERROR: Command errored out with exit status 1: /home/peter/git/lexedata-1.0.0/lenv/bin/python3 /tmp/tmp7dbr261c_in_process.py get_requires_for_build_wheel /tmp/tmpx9nxe50i Check the logs for full command output.
```

It can be something that relates to my system configuration. I also have pyenv installed on my computer to manage Python versions (but no version currently activated). I also performed a global install of lexedata before (sudo pip3 install lexedata), but uninstalled it again.
