### 2022-02-14 Anaphory/lexedata#307 @xrotwang
@peterdekker What does `$ python3 -c"import tempfile;print(tempfile.gettempdir())"
` say? If python thinks that `/tmp` is your temp dir, but it cannot write files to it, then I would assume
> something that relates to my system configuration
