### 2022-02-14 Anaphory/lexedata#309 @peterdekker
JOSS paper: Summary suggestions
https://github.com/openjournals/joss-reviews/issues/4140

I think the Summary section of the JOSS paper (lee linked issue) gives a good overview of why the CLDF format is relevant and that the lexedata library fits in that framework. But I think the summary can be more explicit on what is exactly the functionality that is still missing, and what the toolbox exactly does. The first sentence gives some information on what the toolbox does, but still in quite general terms, this can be further specified.
Of course there is the "Statement of need" section, but I think it would be nice if the reader already gets more of an overview of the goal and functionality of this toolbox in the Summary.
