### 2022-02-14 Anaphory/lexedata#310 @peterdekker
JOSS paper: Statement of need suggestions
https://github.com/openjournals/joss-reviews/issues/4140

I think in this section (analogous to my comment on Summary: https://github.com/Anaphory/lexedata/issues/309), the sense of urgency can be made more clear. What are exactly the tasks researchers working with large computational datasets struggle with, and have not yet been solved using other software packages? How exactly does lexedata help here?
If I understand correctly, the most important innovation of lexedata is that it allows editing the data while keeping the relational structure of the CLDF dataset intact, this can maybe be further stressed and exemplified.

In the first sentence of the section: "Maintaining the integrity as relational database is difficult using general CSV editing tools", I think it would help the reader to make more clear that this "relational database" points to the CLDF format. It is explained in the previous Summary section that the CLDF format is relational, but for a reader not so acquainted with file formats in computational historical/diversity linguistics, it may be good to explain this again.
