### 2022-02-14 Anaphory/lexedata#301 @peterdekker
For me, a reader who is half acquainted with the work and file formats in computational historical linguistics, I think it could be clearer how this software package exactly relates to other software packages in the field. On top of cldfbench that @xrotwang  mentions, how does the functionality of lexedata relate to pycldf, pyclts (which are meant to access datasets) and for example lingpy (performing operations on the data)? Which functionality do they have in common, where do they differ?
