### 2022-02-14 Anaphory/lexedata#311 @peterdekker
JOSS paper: Explain choice for terminal tool
https://github.com/openjournals/joss-reviews/issues/4140

To me, a quite surprising feature of this software package is that it is a terminal tool, not a library used in Python. I think it would be good to explain in the paper why this architecture has been chosen, which benefits does it have for the user?
