### 2022-02-15 Anaphory/lexedata#312 @peterdekker
Similar error for add_concepticon:
```
python -m lexedata.edit.add_concepticon -q -l Name=en --add-concept-set-names --add-definitions
Traceback (most recent call last):
  File "/home/peter/git/lexedata-1.0.0/lenv/lib/python3.9/site-packages/cldfcatalog/config.py", line 67, in get_clone
    return pathlib.Path(self.clones[key])
  File "/usr/lib/python3.9/configparser.py", line 1254, in __getitem__
    raise KeyError(key)
KeyError: 'concepticon'

Traceback (most recent call last):
  File "/usr/lib/python3.9/runpy.py", line 197, in _run_module_as_main
    return _run_code(code, main_globals, None,
  File "/usr/lib/python3.9/runpy.py", line 87, in _run_code
    exec(code, run_globals)
  File "/home/peter/git/lexedata-1.0.0/src/lexedata/edit/add_concepticon.py", line 32, in <module>
    concepticon_path = cldfcatalog.Config.from_file().get_clone("concepticon")
  File "/home/peter/git/lexedata-1.0.0/lenv/lib/python3.9/site-packages/cldfcatalog/config.py", line 69, in get_clone
    raise KeyError('Config {0} has no entry for {1}'.format(self._files, key))
KeyError: 'Config /home/peter/.config/cldf/catalog.ini has no entry for concepticon'
```
