### 2022-02-15 Anaphory/lexedata#312 @peterdekker
And here as well:
```
python -m lexedata.report.segment_inventories -q --language Nzebi # doctest: +NORMALIZE_WHITESPACE
```
