### 2022-02-15 Anaphory/lexedata#313 @peterdekker
edictor error
https://github.com/openjournals/joss-reviews/issues/4140

When following the tour in the docs:
```
python -m lexedata.exporter.edictor --concepts-file concepts_filter
usage: edictor.py [-h] [--metadata METADATA] [--loglevel LOGLEVEL] [-q] [-v] [--languages LANGUAGE [LANGUAGE ...]] [--concepts CONCEPT [CONCEPT ...]] [--cognatesets COGNATESET [COGNATESET ...]] [--output-file OUTPUT_FILE]
edictor.py: error: unrecognized arguments: --concepts-file concepts_filter
```
