### 2022-02-15 Anaphory/lexedata#314 @peterdekker
Docs: Explain why using git is recommended
https://github.com/openjournals/joss-reviews/issues/4140

From the docs, it becomes clear that it is highly recommended to use lexedata in conjunction with git. There is even a page explaining git to beginners. However, I never see the rationale mentioned, why users of lexedata have to use git (e.g. This makes it possible to save intermediate changes in your data/collaborate). I think this could be made more explicit in the documentation (e.g. in the tour or the main page), or possibly even in the paper.
