### 2022-02-15 Anaphory/lexedata#315 @peterdekker
Docs: Explain that tools work on whole directory
https://github.com/openjournals/joss-reviews/issues/4140

From the tour in the docs, it became clear to me that the command line tools mostly do not take a file name as an argument, but work on the whole directory, as this directory represents a CLDF dataset. I think this asks for a bit of a paradigm shift from a user, as many command line tools work on specific files. I think this would be good to mention explicitly in the documentation, for example on the Tour page.
