### 2022-02-15 #4140 @peterdekker
Review:

This paper and software package, lexedata, presents a set of command line tools that can be used to edit datasets in CLDF format, used in computational historical/diversity linguistics, in the CLDF format, maintaining the relational structure of the dataset. I think this is very nice work, as it enables further quantitative research using wordlists in linguistics. The architecture of the software package as a command line tool, and the clear explanation in the documentation, makes this type of dataset editing accessible for researchers without elaborate programming experience. It is also nice how lexedata builds further upon existing data formats, libraries and technologies, such as CLDF, lingpy, git and the command line. The suggestions I have for this work regard mostly clarifications of the paper and the documentation. I only noticed some small issues in the code (https://github.com/Anaphory/lexedata/issues/313 & https://github.com/Anaphory/lexedata/issues/312).

In the paper, I think the urgency and need for this tool, and what it exactly does, can be better described, both in the Summary (https://github.com/Anaphory/lexedata/issues/309) and in the Statement of need (https://github.com/Anaphory/lexedata/issues/310). I think it is important to stress more how this software package relates to other software packages (https://github.com/Anaphory/lexedata/issues/301). It would also be interesting to motivate the choice to create a terminal tool rather than a Python library (https://github.com/Anaphory/lexedata/issues/311).

The documentation tries to make things as clear as possible, with a tour through some useful commands. I think that in the documentation, the usefulness of git for the user can be better explained: https://github.com/Anaphory/lexedata/issues/314 Also, I think it would be good to make clear to the user the paradigm of this set of tools, where command line commands act on a whole dataset/directory, without taking a filename as argument (https://github.com/Anaphory/lexedata/issues/315).

All in all, I suggest acceptance with minor revisions for this paper.
