### 2022-02-15 #4140 @xrotwang
@Bisaloo

# Reviewer report

## Comments

The software and paper fulfill all criteria from the checklist. The software is freely available on GitHub, and additonally distributed via PyPI, includes examples and tests, is tested regularly in multiple configurations via GitHub Actions, is documented on ReadTheDocs and contains a readme, open-source licence and contribution guidelines. Thus, it's a state-of-the-art Python package.

lexedata comes with a laudable amount of documentation that is available online. The CLI also does a good job of guiding users through a sensible data curation workflow. My only nit is that there's no way to list all available commands from the command line (maybe as `python -m lexedata`?) - but maybe reading through the manual is the better start anyway.

### Conclusion

While the CLDF format is slowly taking hold as format for sharing and publishing data in linguistics, this progress is hampered by the fact that most data is still collected and often curated in other formats. Often these "other formats" are some variant of spreadsheets. Thus, the problem that lexedata sets out to solve is real, and so is lexedata's potential. Thus, I recommend acceptance of this paper (ideally - but not necessarily - after the open issues at https://github.com/Anaphory/lexedata/issues are addressed).
