+++ docs/index.rst
@@ -6,7 +6,7 @@ Introduction
 
 Lexedata is a set of tools for managing, editing, and annotating large lexical datasets in :doc:`cross-linguistic data format (CLDF) <cldf>`. You can find how commands are organized in lexedata, and a description of all available commands and operations in the manual. For an example using a simple dataset, see the lexedata tour below.
 
-In order to use lexedata you need to be somewhat familiar with the command line and git. If you are not familiar with CLDF, git and the command line, you can find a brief introduction and further references under "Introductions to the Ecosystem". You can also find a glossary with the main terms that are used in the manual and lexedata help.
+In order to use lexedata you need to be somewhat familiar with the command line and the CLDF format. We also strongly recommend using git for version control, so you can easily access and restore previous versions of your dataset and collaborate with others. If you are not familiar with CLDF, git and the command line, you can find a brief introduction and further references under "Introductions to the Ecosystem". You can also find a glossary with the main terms that are used in the manual and lexedata help.
 
 Lexedata is open access software in development. Please report any problems and suggest any improvements you would like to see by `opening an issue  <https://docs.github.com/en/issues/tracking-your-work-with-issues/creating-an-issue#creating-an-issue-from-a-repository>`_ on the `Lexedata GitHub repository <https://github.com/Anaphory/lexedata/tree/master>`_.
    
