### 2022-02-16 Anaphory/lexedata#314 @nataliacp
I have added a sentence justifying git in the beginning of the tour and in the index file of the documentation. Given the length limit of the paper, i didn't find a very suitable place in there.
