### 2022-02-16 2d2dfe0
added bit on directory vs file and link to cldf
