+++ docs/Manual.md
@@ -8,6 +8,8 @@ You can access the Lexedata tools through commands in your terminal. Every comma
 
 Elements in italics above need to be replaced depending on the exact command you are using. Elements in capital letters need to be replaced depending on the exact operation you want to perform on your dataset. Optional elements are enclosed in brackets. There could be multiple positional arguments, optional arguments and switches (with only a space as a separator).  Positional arguments are not preceded by a hyphen and need to occur in strict order (if there are more than two of them). Optional arguments and switches are always preceded by two hyphens and they can occur in any order after the command. Optional arguments require a value (often there is a default value), while switches do not. Some optional arguments or switches have a short name of one letter in addition to their regular name and in this case they are preceded by one hyphen (e.g. to access the help of any command you can use the switch `--help` or `-h`). Many positional arguments and most optional arguments have default settings.
 
+In order to maintain the integrity of the CLDF format, which is a relational database, lexedata scripts often have to operate on multiple files. It is therefore common that a lexedata command does not have as an argument a single file. Instead, it operates on as many files of the CLDF format as necessary for the operation at hand. For more information on the CLDF format and its files, see [here](https://github.com/Anaphory/lexedata/blob/natalia-review/docs/cldf.md).
+
 Commands in Lexedata are organized in four packages: [lexedata.importer](#importing-data-lexedata-importer) [^1], [lexedata.edit](#editing-a-cldf-dataset-lexedata-edit), [lexedata.report](#reporting-and-checking-data-integrity-lexedata-report), and [lexedata.exporter](#exporting-data-lexedata-exporter) [^1]. If a command name consists of more than one word, the words are separated by an underscore [^1]. However, if optional arguments or switches consist of more than one word, the words are separated by a hyphen. If you need to replace an element in capital letters with a phrase including a space, enclose the whole phrase in quotes (`"MORE THAN ONE WORD"`). 
 
 [^1]: In case you are wondering why the packages are called **importer** and **exporter**, not import and export: The word `import` has special status in Python, so it was not available as name of a package! The hyphen `-` is used as subtraction symbol in Python, so it also may not be part of a name.
