### 2022-02-16 Anaphory/lexedata#315 @nataliacp
I have added a short paragraph in the manual just after the general description of lexedata commands and linked to the CLDF introduction.
