+++ docs/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Lexedata: A Python toolbox to edit CLDF lexical datasets'
+title: 'Lexedata: A toolbox to edit CLDF lexical datasets'
 tags:
   - Python
   - linguistics
