### 2022-02-24 227b57f
clarify metadata argument for commands.
 fixes #315
