### 2022-02-24 Anaphory/lexedata#311 @Anaphory
By your comments, it looks like addressing both #301 and #308 will solve the confusion and explain the design, so we close this and address those.
