### 2022-02-24 Anaphory/lexedata#310 @Anaphory
So we need to make sure we address the shortcomings of the following existing pieces of software:

 - [ ] cldfbench (requires Python skills)
 - [ ] beastling (not flexible in terms of coding, cumbersome subsetting)
 - [ ] edictor (export to and from; not working on native CLDF; difficult for cross-concept cognate sets)
 - [ ] lingpy (expose lexstat with parameters – TBD)
 - [ ] Maybe plain table editing software (no relational checks, difficult to do some comp hist ling manipulations)
 - [ ] actual relational databases (beyond programming skills of most linguists, often difficult to set up)
