### 2022-02-24 Anaphory/lexedata#306 @Anaphory
- [ ] Link to Pypi on the readme and say Lexedata can be installed using `pip` from there.
 - [ ] Mention sphinx and `make html` in the CONTRIBUTING.md for development installs
