### 2022-02-24 91646a9
Re-work some phrases
Fixes #309
Fixes #310
