### 2022-02-24 cdae4a9
Fix typos @xrotwang noticed
There may be more, that's why I asked @nataliacp to review ;)
