+++ docs/paper.md
@@ -108,11 +108,11 @@ from simple terminal commands.
 The extensive lexical dataset editing functionality is currently used by projects
 at UC Berkeley and Universität Zürich for Arawakan and Mawetí-Guaraní languages
 and at Universiteit Gent for Bantu.
+Precursor scripts have also been used for Timor-Alor-Pantar and Austronesian languages [@lexirumah-paper].
 The export to phylogenetic alignments, derived from BEASTling
 [@maurits2017beastling;@beastling14], has been used in different language
-phylogenetics projects that are currently under review
-[@kaiping2019subgrouping;@kaiping2021burst;@gunnink2022bantu]. 
-Precursor scripts have also been used for Timor-Alor-Pantar and Austronesian languages [@lexirumah-paper].
+phylogenetics projects that are currently currently under review
+[@kaiping2019subgrouping;@kaiping2021burst;@gunnink2022bantu].
 
 # Acknowledgement
 Development of Lexedata was funded by the Swiss National Science Foundation
