+++ docs/paper.md
@@ -111,7 +111,7 @@ and at Universiteit Gent for Bantu.
 Precursor scripts have also been used for Timor-Alor-Pantar and Austronesian languages [@lexirumah-paper].
 The export to phylogenetic alignments, derived from BEASTling
 [@maurits2017beastling;@beastling14], has been used in different language
-phylogenetics projects that are currently already under review
+phylogenetics projects that are already under review
 [@kaiping2019subgrouping;@kaiping2021burst;@gunnink2022bantu].
 
 # Acknowledgement
