### 2022-03-03 9abc359
Catch and report when cell value is datetime
Fixes #304
