### 2022-03-03 Anaphory/lexedata#314 @Bisaloo
For future reference, this was fixed in https://github.com/Anaphory/lexedata/commit/c4b9a9b28daa550fbb60f64247dd2e68dde4f6bd.
