### 2022-03-10 62aec0a
Be more explicit about installations
Fixes #306
