### 2022-03-10 Anaphory/lexedata#295 @PaprikaSteiger
If there is no metadata -> error (missing metadata, run add_metadata) "No cldf metadata found, if you have no metadata, export to csv from your excel file and run lexedata.edit.add_metadata"
If
