### 2022-03-10 98d4f2f
Use Set instead of List for XXXOrFromFile
Noticed while working on #313

Also:
Fix #313
