### 2022-03-10 Anaphory/lexedata#312 @Anaphory
To be fixed in the upcoming PR.

 - I'm still confused why you encountered “access 'https://github.com/cldf-clts/clts.git/': Could not resolve host: github.com”, but I don't think I have a way to debug this, and with improved installation instructions and errors, I hope it will not be a problem any more.
 - I fixed the installation instructions, because “No” (`[y/N]`) is *not* a sensible default, contrary to what we had stated. With agreeing to cone the repositories, you would have not run into most of these issues.
 - Added a warning to `add_segments` about installing catalogs. It's the only script that needs a catalog at import-time, so it has a fall-back for when we build documentation on ReadTheDocs – and a normal user should be warned when they run into that. (`report.segment_inventories` imports it, so is thus also handled.)
 - Added a similar warning to `add_concepticon`
 -
