### 2022-03-10 87416c6
Improve catalog handling, when they are missing
Fixes #312
