### 2022-03-16 38e5561
Use zip_longest instead of zip in interleaved form/cogset match
Addresses #327
(To fix that issue, we still need regression tests)
