### 2022-03-17 b884cbc
Fix the actual error in #327
Still needs a regression test
