### 2022-03-17 4623907
Fix test for interleaved with NA forms
