### 2022-03-17 4196d91
Add tests for interleaved import
Also fix a fixture.
