### 2022-03-17 Anaphory/lexedata#332 @Anaphory
Check references in paper
I just noticed a stray `{W3c}` and some ((double brackets)) in the bibliography of the paper. It's probably good if we have another look over the references, @nataliacp.

https://github.com/openjournals/joss-reviews/issues/4140
