+++ src/lexedata/exporter/cognates.py
@@ -529,18 +529,18 @@ if __name__ == "__main__":  # pragma: no cover
     cogsets, judgements = cogsets_and_judgements(
         dataset, args.add_singletons_with_status, args.by_segment, logger
     )
-
-    try:
-        cogset_order = (
-            util.cldf_property(
-                dataset["CognatesetTable", args.sort_cognatesets_by].propertyUrl
+    if args.sort_cognatesets_by:
+        try:
+            cogset_order = (
+                util.cldf_property(
+                    dataset["CognatesetTable", args.sort_cognatesets_by].propertyUrl
+                )
+                or dataset["CognatesetTable", args.sort_cognatesets_by].name
             )
-            or dataset["CognatesetTable", args.sort_cognatesets_by].name
-        )
-    except (KeyError):
-        cli.Exit.INVALID_COLUMN_NAME(
-            f"No column '{args.sort_cognatesets_by}' in your CognatesetTable."
-        )
+        except (KeyError):
+            cli.Exit.INVALID_COLUMN_NAME(
+                f"No column '{args.sort_cognatesets_by}' in your CognatesetTable."
+            )
     sort_cognatesets(cogsets, judgements, cogset_order, size=args.size_sort)
 
     # TODO: wrap the following two blocks into a
