### 2022-03-24 137464e
fixing exporter.cognates
