### 2022-03-25 #4140 @Bisaloo
Oh sorry, I missed your message. I see both reviewers have checked all their review items so we should indeed be good to go.

> Do we then create a new patch version number?

Yes, please and report the new version number here. The reviewed version is listed in the opening comment of this issue.

> upload to PyPI

Yes, I believe it could be a good idea. This is not a blocker for the acceptance though so feel free to do it afterwards if you don't have time right now.

> Zenodo

Yes please, we need a DOI for acceptance. Please archive the new version to Zenodo and report the DOI here. Please note that Zenodo tries to automatically extract metadata from your repo but it can often be wrong. You should double check the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
