+++ src/lexedata/importer/excel_matrix.py
@@ -648,7 +648,7 @@ def excel_parser_from_dialect(
     initialized_cell_parser = getattr(cell_parsers, dialect.cell_parser["name"])(
         output_dataset,
         element_semantics=dialect.cell_parser["cell_parser_semantics"],
-        separation_pattern=fr"([{''.join(dialect.cell_parser['form_separator'])}])",
+        separation_pattern=rf"([{''.join(dialect.cell_parser['form_separator'])}])",
         variant_separator=dialect.cell_parser["variant_separator"],
         add_default_source=dialect.cell_parser.get("add_default_source"),
     )
