+++ src/lexedata/edit/merge_cognate_sets.py
@@ -10,24 +10,18 @@ What other columns give warnings, what other columns give errors?
 
 import argparse
 import typing as t
-from pathlib import Path
 from collections import defaultdict
+from pathlib import Path
 
 import pycldf
 
-from lexedata import cli, util, types
+from lexedata import cli, types, util
+from lexedata.edit.merge_homophones import (Merger, Skip, all_mergers, default,
+                                            first, format_mergers,
+                                            must_be_equal,
+                                            parse_homophones_report,
+                                            parse_merge_override)
 from lexedata.edit.simplify_ids import update_ids
-from lexedata.edit.merge_homophones import (
-    Merger,
-    Skip,
-    all_mergers,
-    default,
-    first,
-    format_mergers,
-    must_be_equal,
-    parse_homophones_report,
-    parse_merge_override,
-)
 
 # TODO: Options given on the command line should have preference over defaults,
 # no matter whether they are given in terms of names ("Parameter_ID") or
@@ -174,6 +168,7 @@ def merge_cogsets(
 
 if __name__ == "__main__":
     parser = cli.parser(
+        __package__ + Path(__file__).stem,
         description="Merge script for cognate sets.",
         epilog="""The default merging functions are:
 {:}
