+++ src/lexedata/report/homophones.py
@@ -79,7 +79,7 @@ def list_homophones(
 
 
 if __name__ == "__main__":
-    parser = cli.parser(__package__ + Path(__file__).stem, description=__doc__)
+    parser = cli.parser(__package__ + "." + Path(__file__).stem, description=__doc__)
     parser.add_argument(
         "--output-file",
         "-o",
