+++ src/lexedata/importer/cognates.py
@@ -133,7 +133,7 @@ def import_cognates_from_excel(
 
 
 if __name__ == "__main__":
-    parser = cli.parser(__package__ + Path(__file__).stem, description=__doc__)
+    parser = cli.parser(__package__ + "." + Path(__file__).stem, description=__doc__)
     parser.add_argument(
         "cogsets",
         nargs="?",
