### 2022-03-31 e3c6ea7
Catch and test case without value column
Fixes #336
