### 2022-04-06 #4140 @Bisaloo
@Anaphory, just to be sure, are you waiting for anything from my end? I have listed the last remaining points before acceptance in https://github.com/openjournals/joss-reviews/issues/4140#issuecomment-1078902419. I also tried to compile the pdf but got confused about which branch contained the current version of the paper.

If you're short on time, no problem. This is definitely not meant to rush you. Feel free to address my comment whenever convenient. There is nothing pressing on my end.
