### 2022-04-06 #4140 @Anaphory
No, I'm just a bit slow the last 10 days or so. There's one issue still open above (https://github.com/Anaphory/lexedata/issues/295) and we haven't got round to fixing it yet, which is holding up the new release, which is holding up the finishing.
