### 2022-04-06 9f0077d
counting rows of interleaved import
