### 2022-04-07 f2721bb
Set – and test – package version in __init__.__version__
