+++ src/lexedata/__init__.py
@@ -10,3 +10,5 @@ on the `Lexedata GitHub repository
 from . import util
 
 __all__ = ["util"]
+
+__version__ = "1.0.0"
