+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = lexedata
-version = 1.0.0
+version = attr: lexedata.__version__
 author = Melvin Steiger, Gereon A. Kaiping
 author_email = gereon.kaiping@geo.uzh.ch
 description = Tools to edit lexical data in CLDF, and transform it to and from other well-known formats.
