### 2022-04-13 4434d8f
Interface of `greedy_modularity_communities` has changed
