### 2022-04-13 5e6be33
Extract and test simple case of the cluster extractor
