### 2022-04-13 fbe5221
Warn about even header rows differently
