+++ src/lexedata/importer/excel_interleaved.py
@@ -46,6 +46,16 @@ def import_interleaved(
             except AttributeError:
                 break
 
+    if ws.max_row % 2 == 0:
+        logger.warning(
+            f"The sheet {ws.title} contained an even number of rows in total "
+            f"({ws.max_row} including headers). "
+            f"Given that all other rows come in pairs (forms+cognatesets), "
+            f"I may have a problem with the format. "
+            f"Are there multiple header rows, "
+            f"or rows that look empty but are not after the end of the word list?"
+        )
+
     for language in cli.tq(
         ws.iter_cols(min_col=2), task="Parsing cells", total=ws.max_column
     ):
