### 2022-04-14 ada4dae
Add warning and info when adding languages
