+++ test/test_additional_import.py
@@ -437,7 +437,6 @@ def test_no_concept_separator(single_import_parameters, caplog):
 
 def test_language_id(single_import_parameters, caplog):
     dataset, target, excel, concept_name = single_import_parameters
-    c_f_language = dataset["FormTable", "languageReference"].name = "Language"
     c_c_id = dataset["ParameterTable", "id"].name
     c_c_name = dataset["ParameterTable", "name"].name
     concepts = {c[c_c_name]: c[c_c_id] for c in dataset["ParameterTable"]}
@@ -445,7 +444,7 @@ def test_language_id(single_import_parameters, caplog):
     sheet = MockSingleExcelSheet(
         [
             [
-                "Language",
+                "Language_ID",
                 "English",
                 "Form",
                 "phonemic",
@@ -458,7 +457,7 @@ def test_language_id(single_import_parameters, caplog):
                 "variants",
             ],
             [
-                "Aché",
+                "ache",
                 "one",
                 "form",
                 "phonemic",
@@ -479,7 +478,7 @@ def test_language_id(single_import_parameters, caplog):
             entries_to_concepts=concepts,
             concept_column=concept_name,
         )
-    assert {"Aché"} == {f[c_f_language] for f in dataset["FormTable"]}
+    assert {"ache"} == {f["Language_ID"] for f in dataset["FormTable"]}
     assert "no LanguageTable" in caplog.text
 
 
