### 2022-04-14 193f49d
Fix test by splitting it in two
