+++ docs/paper.md
@@ -112,7 +112,7 @@ Precursor scripts have also been used for Timor-Alor-Pantar and Austronesian lan
 The export to phylogenetic alignments, derived from BEASTling
 [@maurits2017beastling;@beastling14], has been used in different language
 phylogenetics projects that are already under review
-[@kaiping2019subgrouping;@kaiping2021burst;@gunnink2022bantu].
+[@kaiping2019subgrouping;@gunnink2022bantu].
 
 # Acknowledgement
 Development of Lexedata was funded by the Swiss National Science Foundation
