+++ src/lexedata/__init__.py
@@ -11,4 +11,4 @@ from . import util
 
 __all__ = ["util"]
 
-__version__ = "1.0.0"
+__version__ = "1.0.1"
