+++ CITATION.cff
@@ -11,8 +11,8 @@ authors:
   orcid: "https://orcid.org/0000-0002-5693-975X"
 title: "Lexedata"
 subtitle: "Lexical Data Editing tools"
-version: 1.0.0
+version: 1.0.1
 license: GPL-3.0+
 doi: 10.5281/zenodo.5541167
-date-released: 2021-09-30
+date-released: 2022-04-14
 url: "https://github.com/Anaphory/lexedata"
