### 2022-04-14 #4140 @Anaphory
@Bisaloo
Lexedata version 1.0.1 is now [a tag on the Lexedata repository](https://github.com/Anaphory/lexedata/releases/tag/v1.0.1), [deposited in Zenodo](https://zenodo.org/record/6461055) with DOI 10.5281/zenodo.6461055 (but with the title “Lexedata: Lexical Data Editing tools”, meant to resolve the portmanteau more than giving a functional description as the JOSS title does), uploaded to [PyPI](https://pypi.org/project/lexedata/1.0.1/).

So that will be

> @editorialbot set 10.5281/zenodo.6461055 as archive
> @editorialbot set v1.0.1 as version

for you, I guess? I was just looking for where to specify the archive for the PDF and found those commands.

The paper currently didn't compile last time. The error in https://github.com/openjournals/joss-papers/runs/5690924642?check_suite_focus=true suggests to me to check our bibliography for an entry missing an author, but there is no such entry. Any ideas?
