### 2022-04-14 #4140 @Bisaloo
Everything looks good to me. Could I just ask that you convert [`structure.svg`](https://github.com/Anaphory/lexedata/blob/joss-paper/docs/structure.svg) to `pdf` or `eps` so you can insert a high-quality vector version of the figure instead of the png one please?
