### 2022-04-14 57db0fa
Fix some issues in the bibliography
