+++ docs/bibliography.bib
@@ -7,21 +7,30 @@
   month = jun,
   abstract = {A linguistics-focussed command line tool for generating BEAST XML files},
   copyright = {BSD-2-Clause},
-  keywords = {TAP-Paper,tap-phylo}
+  version={1.5.1},
+  url={https://github.com/lmaurits/BEASTling/}
 }
 
 @inproceedings{cldfbench,
-  title={CLDFBench: Give Your Cross-Linguistic Data a Lift},
-  author={Robert Forkel and Johann-Mattis List},
-  booktitle={LREC},
-  year={2020}
+    title = "{CLDFB}ench: Give Your Cross-Linguistic Data a Lift",
+    author = "Forkel, Robert  and
+      List, Johann-Mattis",
+    booktitle = "Proceedings of the 12th Language Resources and Evaluation Conference",
+    month = may,
+    year = "2020",
+    address = "Marseille, France",
+    publisher = "European Language Resources Association",
+    url = "https://aclanthology.org/2020.lrec-1.864",
+    pages = "6995--7002",
+    abstract = "While the amount of cross-linguistic data is constantly increasing, most datasets produced today and in the past cannot be considered FAIR (findable, accessible, interoperable, and reproducible). To remedy this and to increase the comparability of cross-linguistic resources, it is not enough to set up standards and best practices for data to be collected in the future. We also need consistent workflows for the {``}retro-standardization{''} of data that has been published during the past decades and centuries. With the Cross-Linguistic Data Formats initiative, first standards for cross-linguistic data have been presented and successfully tested. So far, however, CLDF creation was hampered by the fact that it required a considerable degree of computational proficiency. With cldfbench, we introduce a framework for the retro-standardization of legacy data and the curation of new datasets that drastically simplifies the creation of CLDF by providing a consistent, reproducible workflow that rigorously supports version control and long term archiving of research data and code. The framework is distributed in form of a Python package along with usage information and examples for best practice. This study introduces the new framework and illustrates how it can be applied by showing how a resource containing structural and lexical data for Sinitic languages can be efficiently retro-standardized and analyzed.",
+    language = "English",
+    ISBN = "979-10-95546-34-4",
 }
 
 @software{lingpy,
   author       = {Johann-Mattis List and
                   Robert Forkel},
-  title        = {{LingPy. A Python Library for Historical 
-                   Linguistics}},
+  title        = {{LingPy}. A {Python} Library for Historical Linguistics},
   month        = jul,
   year         = 2021,
   publisher    = {Zenodo},
@@ -47,7 +56,7 @@
 }
 
 @misc{cldf11,
-  title = {Cldf/Cldf: {{CLDF}} 1.1.2},
+  title = {{{CLDF}} 1.1.2},
   shorttitle = {Cldf/Cldf},
   author = {Forkel, Robert and Cysouw, Michael and List, Johann-Mattis and Rzymski, Christoph and Greenhill, Simon J. and Moran, Steven},
   year = {2021},
@@ -108,7 +117,7 @@
 }
 
 @article{list2017potential,
-  title = {The {{Potential}} of {{Automatic Word Comparison}} for {{Historical Linguistics}}},
+  title = {The Potential of Automatic Word Comparison for Historical Linguistics},
   author = {List, Johann-Mattis and Greenhill, Simon J. and Gray, Russell D.},
   year = {2017},
   month = jan,
@@ -156,7 +165,7 @@
 @article{gunnink2022bantu,
   author={Gunnink, Hilde and Chousou-Polydouri, Natalia and Bostoen, Koen},
   year = {accepted},
-  title={Divergence and contact in Southern Bantu language and population history: a new phylogeny in cross-disciplinary perspective},
[…]
