+++ docs/bibliography.bib
@@ -149,7 +149,7 @@
 }
 
 @techreport{pollock2015metadata,
-  type = {{W3C} Recommendation},
+  type = {{{W3C}} Recommendation},
   title = {Metadata Vocabulary for Tabular Data},
   author = {Pollock, Rufus and Tennison, Jeni and Kellogg, Gregg and Herman, Ivan},
   year = {2015},
