+++ docs/bibliography.bib
@@ -1,5 +1,5 @@
 
-@misc{beastling14,
+@software{beastling14,
   title = {{{BEASTling}}},
   shorttitle = {A Linguistics-Focussed Command Line Tool for Generating {{BEAST XML}} Files.},
   author = {Maurits, Luke and Forkel, Robert and Kaiping, Gereon A.},
@@ -24,7 +24,6 @@
     pages = "6995--7002",
     abstract = "While the amount of cross-linguistic data is constantly increasing, most datasets produced today and in the past cannot be considered FAIR (findable, accessible, interoperable, and reproducible). To remedy this and to increase the comparability of cross-linguistic resources, it is not enough to set up standards and best practices for data to be collected in the future. We also need consistent workflows for the {``}retro-standardization{''} of data that has been published during the past decades and centuries. With the Cross-Linguistic Data Formats initiative, first standards for cross-linguistic data have been presented and successfully tested. So far, however, CLDF creation was hampered by the fact that it required a considerable degree of computational proficiency. With cldfbench, we introduce a framework for the retro-standardization of legacy data and the curation of new datasets that drastically simplifies the creation of CLDF by providing a consistent, reproducible workflow that rigorously supports version control and long term archiving of research data and code. The framework is distributed in form of a Python package along with usage information and examples for best practice. This study introduces the new framework and illustrates how it can be applied by showing how a resource containing structural and lexical data for Sinitic languages can be efficiently retro-standardized and analyzed.",
     language = "English",
-    ISBN = "979-10-95546-34-4",
 }
 
 @software{lingpy,
