+++ docs/bibliography.bib
@@ -3,7 +3,7 @@
   title = {{{BEASTling}}},
   shorttitle = {A Linguistics-Focussed Command Line Tool for Generating {{BEAST XML}} Files.},
   author = {Maurits, Luke and Forkel, Robert and Kaiping, Gereon A.},
-  year = {2018},
+  year = {2019},
   month = jun,
   abstract = {A linguistics-focussed command line tool for generating BEAST XML files},
   copyright = {BSD-2-Clause},
@@ -148,7 +148,7 @@
 }
 
 @techreport{pollock2015metadata,
-  type = {{{W3C}} Recommendation},
+  type = {Recommendation},
   title = {Metadata Vocabulary for Tabular Data},
   author = {Pollock, Rufus and Tennison, Jeni and Kellogg, Gregg and Herman, Ivan},
   year = {2015},
