### 2022-04-14 #4140 @Anaphory
Done, and also fixed some bits in the bibliography style.
