+++ docs/bibliography.bib
@@ -111,8 +111,8 @@
   pages = {117--125},
   publisher = {{Association for Computational Linguistics}},
   address = {{Stroudsburg, PA, USA}},
-  isbn = {978-1-937284-19-0},
   keywords = {Lexedata,TAP-Paper,tap-phylo},
+  url = {https://aclanthology.org/W12-0216}
 }
 
 @article{list2017potential,
