### 2022-04-14 e7511d9
Open up foreign key datatypes
