### 2022-04-16 3bb1279
Make detect cognates slightly more modular
