### 2022-04-16 e20d302
Add LingPy functions here
So they can be simplified
