+++ src/lexedata/edit/detect_cognates.py
@@ -2,18 +2,20 @@
 
 import csv
 import hashlib
+import itertools
 import typing as t
+from collections import defaultdict
 from pathlib import Path
 
+import lexedata.cli as cli
+import lexedata.types as types
 import lingpy
 import lingpy.compare.partial
 import pycldf
 import pyclts
 import segments
-
-import lexedata.cli as cli
-import lexedata.types as types
 from lexedata.edit.add_segments import bipa
+from lingpy.algorithm import calign, extra
 
 # TODO maybe the CLTS logic from here and there belongs in util?
 
@@ -70,6 +72,221 @@ def filter_function_factory(
         return row["segments"] and row["concept"]
 
     return filter
+
+
+def _charstring(id_, char="X", cls="-"):
+    return "{0}.{1}.{2}".format(id_, char, cls)
+
+
+def get_slices(*args, **kwargs):
+    breakpoint()
+
+
+def get_partial_matrices(
+    self,
+    concept=False,
+    method="sca",
+    scale=0.5,
+    factor=0.3,
+    restricted_chars="_T",
+    mode="global",
+    gop=-2.0,
+    defaults=False,
+    external_scorer=False,  # external scoring function
+    imap_mode=False,
+    sep=lingpy.settings.rcParams["morpheme_separator"],
+    word_sep=lingpy.settings.rcParams["word_separator"],
+    word_seps=lingpy.settings.rcParams["word_separators"],
+    seps=lingpy.settings.rcParams["morpheme_separators"],
+    tones="T",
+    split_on_tones=False,
+):
+    """
+    Function creates matrices for the purpose of partial cognate detection.
+    """
+
+    def function(idxA, idxB, sA, sB):
+        if method == "lexstat":
+            args = [
+                self[idxA, self._numbers][sA[0] : sA[1]],
+                self[idxB, self._numbers][sB[0] : sB[1]],
+                [
+                    self.cscorer[_charstring(self[idxB, self._langid]), n]
+                    for n in self[idxA, self._numbers][sA[0] : sA[1]]
+                ],
+                [
+                    self.cscorer[_charstring(self[idxA, self._langid]), n]
+                    for n in self[idxB, self._numbers][sB[0] : sB[1]]
+                ],
+                self[idxA, self._prostrings][sA[0] : sA[1]],
+                self[idxB, self._prostrings][sB[0] : sB[1]],
+                1,
+                scale,
+                factor,
+                self.cscorer,
+                mode,
+                restricted_chars,
+                1,
+            ]
+        elif method == "sca":
+            args = [
+                [n.split(".", 1)[1] for n in self[idxA, self._numbers][sA[0] : sA[1]]],
+                [n.split(".", 1)[1] for n in self[idxB, self._numbers][sB[0] : sB[1]]],
+                self[idxA, self._weights][sA[0] : sA[1]],
+                self[idxB, self._weights][sB[0] : sB[1]],
+                self[idxA, self._prostrings][sA[0] : sA[1]],
+                self[idxB, self._prostrings][sB[0] : sB[1]],
+                gop,
+                scale,
+                factor,
+                self.rscorer,
+                mode,
+                restricted_chars,
+                1,
+            ]
+        else:
+            raise ValueError(f"Method {method} unknown.")
+        return calign.align_pair(*args)[2]
+
+    concepts = [concept] if concept else sorted(self.rows)
+
+    # we have two basic constraints in the algorithm:
+    # a) set cognacy between morphemes in the same word to zero
+    # b) set cognacy for those parts to zero which are superceded by
+    # another part in all comparisons of two words
+    # essentially, setting things to zero, means setting them to 1, since
+    # we are dealing with distances here
+    for c in concepts:
+
[…]
