### 2022-04-16 b209e6b
Add a get_slices function
