### 2022-04-16 b6bacab
Reduce double negation
