### 2022-04-16 f223261
Move test outside iterated function call
