### 2022-04-16 c75d676
Directly implement LingPy method
