+++ src/lexedata/edit/detect_cognates.py
@@ -25,17 +25,31 @@ def sha1(path):
     return hashlib.sha1(str(path).encode("utf-8")).hexdigest()[:12]
 
 
+def _charstring(id_, char="X", cls="-"):
+    return "{0}.{1}.{2}".format(id_, char, cls)
+
+
+class SimpleScoreDict(dict):
+    def __missing__(self, key):
+        k1, k2 = key
+        if k1 == k2:
+            return 0.0
+        else:
+            return -1.0
+
+
 def clean_segments(segment_string: t.List[str]) -> t.Iterable[pyclts.models.Symbol]:
     """Reduce the row's segments to not contain empty morphemes.
 
     This function removes all unknown sound segments (/0/) from the segments
     string it is passed, and removes empty morphemes by collapsing subsequent
-    morpheme boundary markers (_#◦+→←) into one.
+    morpheme boundary markers (_#◦+→←) into one, normalizing the phonetics in
+    the process.
 
     >>> segments = "t w o _ m o r ph e m e s".split()
     >>> c = clean_segments(segments)
     >>> [str(s) for s in c]
-    ["t", "w", "o", "_", "m", "o", "r", "ph", "e", "m", "e", "s"]
+    ['t', 'w', 'o', '_', 'm', 'o', 'r', 'pʰ', 'e', 'm', 'e', 's']
 
     >>> segments = "+ _ t a + 0 + a t".split()
     >>> c = clean_segments(segments)
@@ -111,35 +125,158 @@ alignment_functions = {
 }
 
 
+def compute_one_matrix(
+    tokens_by_index: t.Mapping[types.Form_ID, t.List[str]],
+    align_function: t.Callable[[types.Form_ID, types.Form_ID, slice, slice], float],
+) -> t.Tuple[
+    t.Mapping[types.Form_ID, t.List[t.Tuple[slice, int]]], t.List[t.List[float]]
+]:
+    """Compute the distance matrix for pairwise alignment of related morphemes.
+
+    Align all pairs of morphemes (presumably each of them part of a form
+    associated to one common concept), while assuming there is no
+    reduplication, so one morpheme in one form can be cognate (and thus have
+    small relative edit distance) to at most one morpheme in another form.
+
+    Return the identifiers of the morphemes, and the corresponding distance
+    matrix.
+
+    The align_function is a function that calculates the alignment score for
+    two slices of token strings, such as
+
+    >>> identity_scorer = SimpleScoreDict()
+    >>> def align(f1, f2): return alignment_functions["global"](
+    ...   f1, f2,
+    ...   [-1 for _ in f1], [-1 for _ in f2],
+    ...   ['X' for _ in f1], ['X' for _ in f2],
+    ...   len(f1), len(f2),
+    ...   scale = 0.5,
+    ...   factor = 0.0,
+    ...   scorer = identity_scorer
+    ...   )[2] * 2 / (len(f2) + len(f1))
+    >>> def slice_align(f1, f2, s1, s2): return align(data[f1][s1], data[f2][s2])
+
+    In the simplest case, this is just a pairwise alignment, with optimum at 0.
+
+    >>> data = {"f1": "form", "f2": "form"}
+    >>> compute_one_matrix(data, slice_align)
+    ({'f1': [(slice(0, 4, None), 0)], 'f2': [(slice(0, 4, None), 1)]}, [[0.0, 0.0], [0.0, 0.0]])
+
+    This goes up with partial matches.
+
+    >>> data = {"f1": "form", "f3": "fowl"}
+    >>> compute_one_matrix(data, slice_align)
[…]
