### 2022-04-16 44405eb
Test partial matrix calculation
