+++ test/test_cognate_detector.py
@@ -1,4 +1,5 @@
 import lingpy
+import numpy
 import pytest
 from lexedata import util
 from lexedata.edit.add_segments import add_segments_to_dataset
@@ -6,11 +7,12 @@ from lexedata.edit.detect_cognates import (
     alignment_functions,
     filter_function_factory,
     get_partial_matrices,
+    partial_cluster,
 )
 
 
 @pytest.fixture(params=list(alignment_functions))
-def aligment_type(request):
+def alignment_type(request):
     return request.param
 
 
@@ -80,23 +82,35 @@ def test_filter_function_factory():
     }
 
 
-def test_partials_compare_lingpy():
-    # Test that our method of computing partial matrices matches the one implemented in LingPy.
-    lex = lingpy.compare.partial.Partial(
+@pytest.fixture
+def lex():
+    return lingpy.compare.partial.Partial(
         {
             0: ["doculect", "concept", "tokens"],
             1: ["l1", "c1", list("form")],
             2: ["l2", "c1", list("folm")],
             3: ["l3", "c1", list("furm+forn")],
             4: ["l2", "c1", list("diff't")],
+            5: ["l1", "c2", list("room")],
+            6: ["l2", "c2", list("loom")],
+            7: ["l3", "c2", list("ruum+roon")],
+            8: ["l2", "c2", list("other")],
+            9: ["l1", "c1", list("dorm")],
+            10: ["l2", "c1", list("dolm")],
+            11: ["l3", "c1", list("durm+dorn")],
+            12: ["l2", "c1", list("unrelated")],
         }
     )
+
+
+def test_partial_matrices_compare_lingpy(lex, alignment_type):
+    # Test that our method of computing partial matrices matches the one implemented in LingPy.
     lex.get_scorer(runs=10000, ratio=(1, 0), threshold=0.7)
     lingpy_matrices = {
         c: matrix
         for c, mapping, matrix in lex._get_partial_matrices(
             method="lexstat",
-            mode="global",
+            mode=alignment_type,
             imap_mode=True,
         )
     }
@@ -104,9 +118,29 @@ def test_partials_compare_lingpy():
         c: matrix
         for c, mapping, matrix in get_partial_matrices(
             lex,
-            ["c1"],
+            ["c1", "c2"],
             method="lexstat",
-            mode="global",
+            mode=alignment_type,
         )
     }
+    for concept in set(lingpy_matrices) | set(lexedata_matrices):
+        assert numpy.allclose(
+            lingpy_matrices[concept], lexedata_matrices[concept], equal_nan=True
+        )
+
+
+def test_partial_cluster_compare_lingpy(lex, alignment_type):
+    # Test that our method of computing partial matrices matches the one implemented in LingPy.
+    lex.get_scorer(runs=10000, ratio=(1, 0), threshold=0.7)
+    lingpy_matrices = lex.partial_cluster(
+        method="lexstat",
+        mode=alignment_type,
+        imap_mode=True,
+    )
+    lexedata_matrices = partial_cluster(
+        lex,
+        method="lexstat",
+        mode=alignment_type,
+    )
+
     assert lingpy_matrices == lexedata_matrices
