### 2022-04-17 595ecf8
Implement complete fix, and tests
Fixes #349
