### 2022-04-19 #4140 @Bisaloo
@Anaphory @nataliacp, it looks like references should only have a single ISSN. Could you please have a look?
