+++ docs/bibliography.bib
@@ -180,7 +180,7 @@
   volume = {112},
   number = {43},
   pages = {13296--13301},
-  issn = {0027-8424, 1091-6490},
+  issn = {1091-6490},
   doi = {10.1073/pnas.1503793112},
   abstract = {Unlike most other biological species, humans can use cultural innovations to occupy a range of environments, raising the intriguing question of whether human migrations move relatively independently of habitat or show preferences for familiar ones. The Bantu expansion that swept out of West Central Africa beginning {$\sim$}5,000 y ago is one of the most influential cultural events of its kind, eventually spreading over a vast geographical area a new way of life in which farming played an increasingly important role. We use a new dated phylogeny of {$\sim$}400 Bantu languages to show that migrating Bantu-speaking populations did not expand from their ancestral homeland in a ``random walk'' but, rather, followed emerging savannah corridors, with rainforest habitats repeatedly imposing temporal barriers to movement. When populations did move from savannah into rainforest, rates of migration were slowed, delaying the occupation of the rainforest by on average 300 y, compared with similar migratory movements exclusively within savannah or within rainforest by established rainforest populations. Despite unmatched abilities to produce innovations culturally, unfamiliar habitats significantly alter the route and pace of human dispersals.},
   langid = {english},
@@ -211,7 +211,7 @@
   volume = {323},
   number = {5913},
   pages = {479--483},
-  issn = {0036-8075, 1095-9203},
+  issn = {1095-9203},
   doi = {10.1126/science.1166858},
   abstract = {Debates about human prehistory often center on the role that population expansions play in shaping biological and cultural diversity. Hypotheses on the origin of the Austronesian settlers of the Pacific are divided between a recent ``pulse-pause'' expansion from Taiwan and an older ``slow-boat'' diffusion from Wallacea. We used lexical data and Bayesian phylogenetic methods to construct a phylogeny of 400 languages. In agreement with the pulse-pause scenario, the language trees place the Austronesian origin in Taiwan approximately 5230 years ago and reveal a series of settlement pauses and expansion pulses linked to technological and social innovations. These results are robust to assumptions about the rooting and calibration of the trees and demonstrate the combined power of linguistic scholarship, database technologies, and computational phylogenetic methods for resolving questions about human prehistory. Geographical distribution of Austronesian languages reveals the rate and direction of migrations from Taiwan 5200 years ago. Geographical distribution of Austronesian languages reveals the rate and direction of migrations from Taiwan 5200 years ago.},
   copyright = {American Association for the Advancement of Science},
